main/pyproject.toml
[tool.poetry]
name = "platform"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
grpcio = "^1.66.1"
grpcio-tools = "^1.66.1"
protobuf = "^5.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
